Added random entry command for the featured event
The event can be configured. All messages are configurable, although
perhaps only the case when an entry has been found would be enough.

Added default template strings for random command

Note that the default template for the found entry is a bit clumsy. This
is because the API does not provide an URL. The URL should not be
hardcoded to alakajam.com either. This could be circumvented either by
adding a configuration option for the entry URL structure or by
modifying the API. Too bad YAML does not support string concatenation
directly.

Added entry title & authors to random entry string